Fix double to int

+++ src/main/java/ru/mishaneyt/protectionstone/util/serial/LocationSerialization.java
@@ -23,9 +23,9 @@ public class LocationSerialization {
   public static @NotNull Location deserialize(final @NotNull Map<String, Object> data) {
     var worldName = (String) data.get("world");
     var world = Bukkit.getWorld(worldName);
-    var x = (double) data.get("x");
-    var y = (double) data.get("y");
-    var z = (double) data.get("z");
+    var x = (int) data.get("x");
+    var y = (int) data.get("y");
+    var z = (int) data.get("z");
     return new Location(world, x, y, z);
   }
 }
